lib/backend/requirements.txt
fastapi
uvicorn
psutil
requests
pydantic[email]
PyJWT[crypto]
email-validator
